Added a simple clock!

+++ src/freelink.d
@@ -1,14 +1,9 @@
 module freelink;
-import computer, file, nls;
-import std.stdio, std.file;
-version(Windows) import std.c.windows.windows: Sleep;
-import std.thread: Thread;
-import contrib.SDL, gui.base, gui.frame, gui.text, xml;
+import std.file, std.stdio, std.thread, std.c.time: sleep;
+import SDL, gui.all, nls, file, computer, xml;
 
-import tools.iter;
-bool writeOn(ref wchar[] target, ref size_t offset, wchar[][] text...) {
+bool writeOn(wchar[] target, ref size_t offset, wchar[][] text...) {
   if (offset==size_t.max) {
-    offset=0;
     return false; /// Ensure the final newline
   }
   wchar[] str=(iterate(text)~reduces!("_~=__"))[offset..$];
@@ -26,8 +21,8 @@ bool writeOn(ref wchar[] target, ref size_t offset, wchar[][] text...) {
 TextGenerator WriteGrid(wchar[] text) {
   struct holder {
     wchar[] text;
-    size_t offset=0;
-    bool call(ref wchar[] target, bool reset) { return writeOn(target, offset, text); }
+    size_t offset;
+    bool call(wchar[] target, bool reset) { if (reset) offset=0; return writeOn(target, offset, text); }
   }
   auto foo=new holder; foo.text=text; return &foo.call;
 }
@@ -37,7 +32,11 @@ bool between(T)(T v, T lower, T upper) { return (v>=lower)&&(v<upper); }
 import std.date:getUTCtime;
 class Prompt {
   long ms;
-  wchar[] inbuffer; wchar[] line; size_t offset; void delegate(wchar[]) lineCB;
+  wchar[] inbuffer; /// typing buffer
+  wchar[] line; /// current line
+  size_t line_offset;
+  size_t curpos; /// cursor position
+  void delegate(wchar[]) lineCB;
   bool show=false;
   /// Process the input buffer
   void push() {
@@ -45,21 +44,25 @@ class Prompt {
     if(show) {
       while (inbuffer.length) {
         switch (inbuffer[0]) {
-          case 8: if (line.length) line=line[0..$-1]; break; /// backspace
-          case 13: lineCB(line); line=""; break; /// CR
-          default: line~=inbuffer[0];
+          case 8: if (line.length) line=(curpos?line[0..curpos-1]:"")~line[curpos..$]; if (curpos) --curpos; break; /// backspace
+          case 13: lineCB(line); line=""; curpos=0; break; /// CR
+          case 275: if (curpos<line.length) ++curpos; break; /// arrow right
+          case 276: if (curpos) --curpos; break; /// arrow left
+          default: line=line[0..curpos]~inbuffer[0]~line[curpos..$]; ++curpos; break;
         }
         inbuffer=inbuffer[1..$];
         if (!show) break;
       }
     }
   }
-  bool generate(ref wchar[] target, bool reset) {
+  bool generate(wchar[] target, bool reset) {
     assert(target.length>2, "Text field width too small to be usable any more; must be >2");
     bool blink=((getUTCtime/ms)%2)?true:false;
-    if (reset) offset=0;
-    if (show) return writeOn(target, offset, "> "w, line, blink?"_"w:" "w);
-    else return writeOn(target, offset, blink?"_"w:" "w);
+    if (reset) line_offset=0;
+    auto line2=line.dup~" "w;
+    if (blink) line2[curpos]='_';
+    if (show) return writeOn(target, line_offset, "> "w, line2);
+    else return writeOn(target, line_offset, line2);
   }
   TextGenerator toStatic() { return WriteGrid("> "~line); }
   void handle(SDL_keysym sym) {
@@ -67,7 +70,7 @@ class Prompt {
       inbuffer~=sym.unicode;
       writefln("Added character ", cast(ubyte[])[sym.unicode], " == ", sym.unicode);
     } else switch (cast(int)sym.sym) {
-      case 8, 13: inbuffer~=cast(char)sym.sym;
+      case 8, 13, 275, 276: inbuffer~=cast(wchar)sym.sym; break;
       default: writefln("Strange sym: ", sym.sym, " which is '", sym.unicode, "'");
     }
   }
@@ -79,7 +82,21 @@ class Prompt {
 
 void delegate(SDL_keysym) KeyHandler=null;
 
-import tools.threads, tools.ext, std.utf, std.string, std.stdio: format;
+import std.utf, std.traits: isStaticArray;
+wchar[] mysformat(T)(T t) {
+  static if(is(T: wchar[])) return t; else
+  static if(isStaticArray!(T)) return mysformat(t[]); else
+  static if(is(T: char[])) return t.toUTF16(); else
+  static assert(false, T.stringof~" not supported (yet)!");
+}
+
+wchar[] myformat(T...)(T t) {
+  wchar[] res;
+  foreach (v; t) res~=mysformat(v);
+  return res;
+}
+
+import tools.threads, tools.ext, std.string, std.stdio: format;
 class TTY {
   Prompt p;
   Font.GridTextField field;
@@ -109,27 +126,38 @@ class TTY {
       gens ~= gens[$-1];
       gens[$-2] = p.toStatic;
     }
+    p.line="";
     return res;
-    
   }
-  void write(wchar[] t) {
+  void writef(wchar[] t) {
     scope(failure) writefln("!write");
     dtb in OSThread;
-    if (p.show) throw new Exception("Trying to print to console while waiting for input: your threads are messed up");
+    p.line~=t;
+  }
+  void writegen(TextGenerator tg) {
+    dtb in OSThread;
+    if (p.line.length) throw new Exception("Trying to write generator on used line ("~p.line.toUTF8()~") - text would be overwritten!");
     with (field) {
       gens ~= gens[$-1];
-      gens[$-2] = WriteGrid(t);
+      gens[$-2] = tg;
     }
+    p.line="";
   }
-  void writefln(T...)(T t) { write(format("", replace(t, "%", "%%")).toUTF16()); }
+  void newline() {
+    dtb in OSThread;
+    with (field) {
+      gens ~= gens[$-1];
+      gens[$-2] = WriteGrid(p.line);
+    }
+    p.line="";
+  }
+  void writefln(T...)(T t) { writef(myformat(t).toUTF16()); newline; }
   void push() { p.push; }
 }
 
-import png;
-import std.c.time: sleep;
 import tools.threadpool, tools.ext;
 import std.process;
-static import std.stream;
+static import std.date, std.stream;
 void main ()
 {
   auto pool=new ThreadPool(1);
@@ -160,7 +188,22 @@ void main ()
   pool.addTask({
     while (true) {
       auto inp=t.readln;
-      system((inp~" 2>>tmp >> tmp").toUTF8()); // "
+      if (inp=="cycle") {
+        foreach (x; Integers[0..10]) {
+          t.writef(".");
+          system("sleep 1");
+        }
+        t.writefln();
+        continue;
+      }
+      if (inp=="date") {
+        t.writegen((wchar[] target, bool reset, ref size_t offset, ref wchar[] datestr) {
+          if (reset) { offset=0; datestr=std.date.toString(std.date.getUTCtime()).toUTF16(); }
+          return writeOn(target, offset, datestr);
+        } ~ rfix(0, ""w));
+        continue;
+      }
+      system((inp~" 2>>tmp >> tmp").toUTF8());
       foreach (result; (cast(char[])read("tmp")).split("\n")) t.writefln(result);
       system("rm tmp"); 
     }
